Fix ignored paths
